Add "promises" and "tabs" commands.
The tabs command just list all the tabs, a baby step towards the promises
command.

The promises command uses a required url filter to decide what tabs you
care about.  When it finds a matching one, it lists the unresolved
promises and their call stacks.  It does not currently leverage the
dependent promises list, but it could.

The promises command also lists workers associated with the tab, but it
seems to be the case that there's currently no way to cause a Promises
actor to be instantiated for the worker (and it also may not work for
that case anyways, although the non-devtools platform stuff does seem
like there was forethought in terms of supporting it.)